requirements.txt
langchain==0.1.20
langchain-community==0.0.30
langchain-core==0.1.40
python-dotenv
pypdf
pymupdf
wikipedia
langchain_text_splitters
langchain-openai
sentence_transformers
langchain_huggingface
langchain_huggingface
faiss-cpu
langchain_chroma
langchain_groq
streamlit